attempt to work on docker on windows
multiple users have attempted to run influxdb in a docker container
with a windows host and a volume mounted from windows. that causes
problems because it apparently uses samba/cifs which does not
support fsync on directories. this patchset will, once, at startup,
check if it can fsync "/". if it receives an EINVAL as is what
appears to happen on samba/cifs, then it will not attempt to fsync
any more. this should help.

updates #9833.
updates #9630.